Drop redundant/outdated copy of the script `rrsync`
The script `rrsync` was a copy of an upstream script back in the time.
An updated version of this script is provided by the package `rsync`
for years (roughly one week after the latest build of this repository).
Thus it is no longer helpful to copy it, quite the contrary.